requirements.txt
requests
wikipedia-api
Pillow
psutil